fix detection of subs coming from roles
Sadly we do need use the Moose or Mouse MOP as applicable to get the
"official" method list.  A sub from a role is the same ref, so it is still
labelled with the original class name, so we can inspect that class to see if
it is a Role::Tiny.